serverless/comparisons/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
boto3==1.35.10
botocore==1.35.10
certifi==2024.8.30
deprecation==2.1.0
frozenlist==1.4.1
gotrue==2.7.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
iniconfig==2.0.0
jmespath==1.0.1
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
postgrest==0.16.11
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.2
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.3
supafunc==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
yarl==1.9.6
